asdf

+++ torch_baselines/DQN/dqn.py
@@ -109,11 +109,11 @@ class DQN:
     def _train_step(self, step, learning_rate):
         # Sample a batch from the replay buffer
         data = self.replay_buffer.sample(self.batch_size)
-        obses = [torch.from_numpy(o) for o in data[0]]
+        obses = [torch.from_numpy(o).float() for o in data[0]]
         obses = [o.permute(0,3,1,2) if len(o.shape) == 4 else o for o in obses]
         actions = torch.from_numpy(data[1])
         rewards = torch.from_numpy(data[2])
-        nxtobses = [torch.from_numpy(o) for o in data[3]]
+        nxtobses = [torch.from_numpy(o).float() for o in data[3]]
         nxtobses = [no.permute(0,3,1,2) if len(no.shape) == 4 else no for no in nxtobses]
         dones = (~torch.from_numpy(data[4])).float()
         vals = self.model(obses).gather(-1,actions)
